cli: Use new devnet stake-pool program id (#172)
* cli: Use devnet stake-pool program id

#### Problem

The devnet deployment of the stake pool program has been out of date for
a long time because the upgrade key has been lost.

#### Summary of changes

Since the key is lost, we need to use a different program id for devnet.
Add that key and use it properly in the CLI. This PR also includes a
change to specify the a program id to use.

* Add devnet as an acceptable word

* Parametrize get_stake_pools to take program id